pyproject.toml
[tool.poetry]
name = "lambo"
version = "0.1.0"
description = ""
authors = ["HKGx <19597269+HKGx@users.noreply.github.com>"]

[tool.poetry.scripts]
lambo = 'lambo.main:main'
migrate = "lambo.migrations:main"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = { extras = ["dotenv"], version = "^1.9.2" }
'discord.py' = { git = "https://github.com/Rapptz/discord.py.git" }
psycopg = { extras = ["binary"], version = "^3.0.16" }

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
autopep8 = "^1.7.0"
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
